+++ rust/crd/src/lib.rs
@@ -1,14 +1,18 @@
 use serde::{Deserialize, Serialize};
+use snafu::{ResultExt, Snafu};
+use stackable_operator::client::Client;
+use stackable_operator::commons::s3::{InlinedS3BucketSpec, S3BucketDef, S3ConnectionSpec};
+use stackable_operator::kube::ResourceExt;
 use stackable_operator::{
+    commons::{opa::OpaConfig, s3::S3ConnectionDef},
     kube::CustomResource,
-    opa::OpaConfig,
     product_config_utils::{ConfigError, Configuration},
     role_utils::Role,
     schemars::{self, JsonSchema},
 };
 use std::collections::BTreeMap;
 use std::str::FromStr;
-use strum::{Display, EnumIter, EnumString};
+use strum::{Display, EnumDiscriminants, EnumIter, EnumString, IntoStaticStr};
 
 pub const APP_NAME: &str = "druid";
 
@@ -65,6 +69,29 @@ pub const CREDENTIALS_SECRET_PROPERTY: &str = "credentialsSecret";
 pub const PROMETHEUS_PORT: &str = "druid.emitter.prometheus.port";
 pub const DRUID_METRICS_PORT: u16 = 9090;
 
+// container locations
+pub const S3_SECRET_DIR_NAME: &str = "/stackable/secrets";
+const ENV_S3_ACCESS_KEY: &str = "AWS_ACCESS_KEY_ID";
+const ENV_S3_SECRET_KEY: &str = "AWS_SECRET_ACCESS_KEY";
+const SECRET_KEY_S3_ACCESS_KEY: &str = "accessKey";
+const SECRET_KEY_S3_SECRET_KEY: &str = "secretKey";
+
+#[derive(Snafu, Debug, EnumDiscriminants)]
+#[strum_discriminants(derive(IntoStaticStr))]
+#[allow(clippy::enum_variant_names)]
+pub enum Error {
+    #[snafu(display("failed to resolve S3 connection"))]
+    ResolveS3Connection {
+        source: stackable_operator::error::Error,
+    },
+    #[snafu(display("failed to resolve S3 bucket"))]
+    ResolveS3Bucket {
+        source: stackable_operator::error::Error,
+    },
+    #[snafu(display("2 differing s3 connections were given, this is unsupported by Druid"))]
+    IncompatibleS3Connections,
+}
+
 #[derive(Clone, CustomResource, Debug, Deserialize, JsonSchema, Serialize)]
 #[kube(
     group = "druid.stackable.tech",
@@ -94,7 +121,7 @@ pub struct DruidClusterSpec {
     pub routers: Role<DruidConfig>,
     pub metadata_storage_database: DatabaseConnectionSpec,
     pub deep_storage: DeepStorageSpec,
-    pub s3: Option<S3Spec>,
+    pub ingestion: Option<IngestionSpec>,
     pub zookeeper_config_map_name: String,
     pub opa: Option<OpaConfig>,
 }
@@ -150,11 +177,32 @@ impl DruidRole {
     }
 
     /// Returns the start commands for the different server types.
-    pub fn get_command(&self, _version: &str) -> Vec<String> {
+    pub fn get_command(&self, mount_s3_credentials: bool) -> Vec<String> {
+        let mut shell_cmd = vec![];
+        if mount_s3_credentials {
+            shell_cmd.push(format!(
+                "export {env_var}=$(cat {secret_dir}/{file_name})",
+                env_var = ENV_S3_ACCESS_KEY,
+                secret_dir = S3_SECRET_DIR_NAME,
+                file_name = SECRET_KEY_S3_ACCESS_KEY
+            ));
+            shell_cmd.push(format!(
+                "export {env_var}=$(cat {secret_dir}/{file_name})",
+                env_var = ENV_S3_SECRET_KEY,
+                secret_dir = S3_SECRET_DIR_NAME,
+                file_name = SECRET_KEY_S3_SECRET_KEY
+            ));
+        }
+        shell_cmd.push(format!(
+            "{} {} {}",
+            "/stackable/druid/bin/run-druid",
+            self.get_process_name(),
+            "/stackable/conf",
+        ));
         vec![
-            "/stackable/druid/bin/run-druid".to_string(),
-            self.get_process_name().to_string(),
-            "/stackable/conf".to_string(),
+            "/bin/sh".to_string(),
+            "-c".to_string(),
+            shell_cmd.join(" && "),
         ]
     }
 }
@@ -183,6 +231,72 @@ impl DruidCluster {
             self.role_service_name(role)?,
             self.metadata.namespace.as_ref()?
         ))
+    }
+
+    /// If an s3 connection for ingestion is given, as well as an s3 connection for deep storage, they need to be the same.
+    /// This function returns the resolved connection, or raises an Error if the connections are not identical.
+    pub async fn get_s3_connection(
+        &self,
+        client: &Client,
+    ) -> Result<Option<S3ConnectionSpec>, Error> {
+        // retrieve connection for ingestion (can be None)
+        let ingestion_conn = if let Some(ic) = self
+            .spec
+            .ingestion
+            .as_ref()
+            .and_then(|is| is.s3connection.as_ref())
+        {
+            Some(
+                ic.resolve(client, self.namespace().as_deref())
+                    .await
+                    .context(ResolveS3ConnectionSnafu)?,
+            )
+        } else {
+            None
+        };
+
+        // retrieve connection for deep storage (can be None)
+        let storage_conn = match &self.spec.deep_storage {
+            DeepStorageSpec::S3(s3_spec) => {
+                let inlined_bucket: InlinedS3BucketSpec = s3_spec
+                    .bucket
+                    .resolve(client, self.namespace().as_deref())
+                    .await
+                    .context(ResolveS3BucketSnafu)?;
+                inlined_bucket.connection
+            }
+            _ => None,
+        };
+
+        // if both connections are specified and are identical, return it
+        // if they differ, raise an error
+        // if only one connection is specified, return it
+        if ingestion_conn.is_some() && storage_conn.is_some() {
+            if ingestion_conn == storage_conn {
+                Ok(ingestion_conn)
+            } else {
+                Err(Error::IncompatibleS3Connections)
+            }
+        } else if ingestion_conn.is_some() {
+            Ok(ingestion_conn)
+        } else if storage_conn.is_some() {
+            Ok(storage_conn)
+        } else {
+            Ok(None)
+        }
+    }
+
+    /// Returns true if the cluster uses an s3 connection.
+    /// This is a quicker convenience function over the [DruidCluster::get_s3_connection] function.
+    pub fn uses_s3(&self) -> bool {
+        let s3_ingestion = self
+            .spec
+            .ingestion
+            .as_ref()
+            .and_then(|spec| spec.s3connection.as_ref())
+            .is_some();
+        let s3_storage = self.spec.deep_storage.is_s3();
+        s3_ingestion || s3_storage
     }
 }
 
@@ -218,38 +332,43 @@ impl Default for DbType {
     }
 }
 
-#[derive(Clone, Debug, Deserialize, Eq, JsonSchema, PartialEq, Serialize, Display, EnumString)]
-pub enum DeepStorageType {
+#[derive(Clone, Debug, Deserialize, JsonSchema, PartialEq, Serialize, Display)]
+#[serde(rename_all = "camelCase")]
+pub enum DeepStorageSpec {
     #[serde(rename = "hdfs")]
     #[strum(serialize = "hdfs")]
-    HDFS,
-
-    #[serde(rename = "s3")]
+    HDFS(HdfsDeepStorageSpec),
     #[strum(serialize = "s3")]
-    S3,
+    S3(S3DeepStorageSpec),
 }
 
-impl Default for DeepStorageType {
-    fn default() -> Self {
-        Self::HDFS
+impl DeepStorageSpec {
+    pub fn is_hdfs(&self) -> bool {
+        matches!(self, DeepStorageSpec::HDFS(_))
+    }
+
+    pub fn is_s3(&self) -> bool {
+        matches!(self, DeepStorageSpec::S3(_))
     }
 }
 
-#[derive(Clone, Debug, Default, Deserialize, JsonSchema, Serialize)]
+#[derive(Clone, Debug, Deserialize, JsonSchema, PartialEq, Serialize)]
 #[serde(rename_all = "camelCase")]
-pub struct DeepStorageSpec {
-    pub storage_type: DeepStorageType,
-    pub storage_directory: Option<String>,
-    // S3 only
-    pub bucket: Option<String>,
+pub struct HdfsDeepStorageSpec {
+    pub storage_directory: String,
+}
+
+#[derive(Clone, Debug, Deserialize, JsonSchema, PartialEq, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct S3DeepStorageSpec {
+    pub bucket: S3BucketDef,
     pub base_key: Option<String>,
 }
 
 #[derive(Clone, Debug, Default, Deserialize, JsonSchema, Serialize)]
 #[serde(rename_all = "camelCase")]
-pub struct S3Spec {
-    pub credentials_secret: String,
-    pub endpoint: Option<String>,
+pub struct IngestionSpec {
+    pub s3connection: Option<S3ConnectionDef>,
 }
 
 #[derive(Clone, Debug, Deserialize, Eq, JsonSchema, PartialEq, Serialize, Default)]
@@ -261,18 +380,11 @@ impl Configuration for DruidConfig {
 
     fn compute_env(
         &self,
-        resource: &Self::Configurable,
+        _resource: &Self::Configurable,
         _role_name: &str,
     ) -> Result<BTreeMap<String, Option<String>>, ConfigError> {
-        let mut result = BTreeMap::new();
-        // s3
-        if let Some(s3) = &resource.spec.s3 {
-            result.insert(
-                CREDENTIALS_SECRET_PROPERTY.to_string(),
-                Some(s3.credentials_secret.clone()),
-            );
-        }
-        Ok(result)
+        let mut _result = BTreeMap::new();
+        Ok(_result)
     }
 
     fn compute_cli(
@@ -330,11 +442,8 @@ impl Configuration for DruidConfig {
                     result.insert(MD_ST_PASSWORD.to_string(), Some(password.to_string()));
                 }
                 // s3
-                if let Some(s3) = &resource.spec.s3 {
-                    if let Some(endpoint) = &s3.endpoint {
-                        result.insert(S3_ENDPOINT_URL.to_string(), Some(endpoint.to_string()));
-                        extensions.push(EXT_S3.to_string());
-                    }
+                if resource.uses_s3() {
+                    extensions.push(EXT_S3.to_string());
                 }
                 // OPA
                 if let Some(_opa) = &resource.spec.opa {
@@ -349,17 +458,24 @@ impl Configuration for DruidConfig {
                     // The opaUri still needs to be set, but that requires a discovery config map and is handled in the druid_controller.rs
                 }
                 // deep storage
-                let ds = &resource.spec.deep_storage;
-                result.insert(DS_TYPE.to_string(), Some(ds.storage_type.to_string()));
-                if let Some(bucket) = &ds.bucket {
-                    result.insert(DS_BUCKET.to_string(), Some(bucket.to_string()));
-                }
-                if let Some(key) = &ds.base_key {
-                    result.insert(DS_BASE_KEY.to_string(), Some(key.to_string()));
-                }
-                // TODO resolve endpoint from namenode name...
-                if let Some(dir) = &ds.storage_directory {
-                    result.insert(DS_DIRECTORY.to_string(), Some(dir.to_string()));
+                result.insert(
+                    DS_TYPE.to_string(),
+                    Some(resource.spec.deep_storage.to_string()),
+                );
+                match &resource.spec.deep_storage {
+                    DeepStorageSpec::HDFS(hdfs_spec) => {
+                        result.insert(
+                            DS_DIRECTORY.to_string(),
+                            Some(hdfs_spec.storage_directory.clone()),
+                        );
+                    }
+                    DeepStorageSpec::S3(s3_spec) => {
+                        if let Some(key) = &s3_spec.base_key {
+                            result.insert(DS_BASE_KEY.to_string(), Some(key.to_string()));
+                        }
+                        // bucket information (name, connection) needs to be resolved first,
+                        // that is done directly in the controller
+                    }
                 }
                 // other
                 result.insert(
@@ -395,6 +511,7 @@ fn build_string_list(strings: &[String]) -> String {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::DeepStorageSpec::HDFS;
     use stackable_operator::role_utils::CommonConfiguration;
     use stackable_operator::role_utils::RoleGroup;
     use std::collections::HashMap;
@@ -466,8 +583,10 @@ mod tests {
                     .collect::<HashMap<_, _>>(),
                 },
                 metadata_storage_database: Default::default(),
-                deep_storage: Default::default(),
-                s3: None,
+                deep_storage: HDFS(HdfsDeepStorageSpec {
+                    storage_directory: "/path/to/dir".to_string(),
+                }),
+                ingestion: Default::default(),
                 zookeeper_config_map_name: Default::default(),
                 opa: Default::default(),
             },