+++ rust/operator-binary/src/main.rs
@@ -4,7 +4,7 @@ mod druid_controller;
 
 use clap::Parser;
 use futures::StreamExt;
-use stackable_druid_crd::DruidCluster;
+use stackable_druid_crd::{DruidCluster, APP_NAME};
 use stackable_operator::{
     cli::Command,
     cli::ProductOperatorRun,
@@ -33,14 +33,13 @@ struct Opts {
 
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
-    stackable_operator::logging::initialize_logging("DRUID_OPERATOR_LOG");
-
     let opts = Opts::parse();
     match opts.cmd {
         Command::Crd => println!("{}", serde_yaml::to_string(&DruidCluster::crd())?),
         Command::Run(ProductOperatorRun {
             product_config,
             watch_namespace,
+            tracing_target,
         }) => {
             stackable_operator::utils::print_startup_string(
                 built_info::PKG_DESCRIPTION,
@@ -49,6 +48,11 @@ async fn main() -> anyhow::Result<()> {
                 built_info::TARGET,
                 built_info::BUILT_TIME_UTC,
                 built_info::RUSTC_VERSION,
+            );
+            stackable_operator::logging::initialize_logging(
+                "DRUID_OPERATOR_LOG",
+                APP_NAME,
+                tracing_target,
             );
             let product_config = product_config.load(&[
                 "deploy/config-spec/properties.yaml",